upload birdvox_pcen_cnn_adaptive-threshold-T1800
this model is trained on BirdVox-70k with units 01, 02, 03 as training and 05 and 07 as test. There are 8 geometrical augmentations. The size of training and validation set is of the order of 500k examples.
A model trained on BirdVox-300k with 15% hold out validation (class conditional and augmentation conditional but not sensor conditional) is forthcoming. It will have 30 geometrical augmentations, hence a total of about 8M examples.